Added support for using file that specify paths to files to process
A new optional argument "i" was created and added to program. When this
argument is used, the program will process the input file as if it
contains a list of files to process.

For example, if ravioli run with the following command

ravioli -i in.txt

where the contents of ravioli is the following

in.txt
    ./test.c
    ./test2.c
    ../../test3.c

ravioli will read the contents of in.txt internally, then process the
each file.

the -i command is similar to specifying a directory, bunt instead of
having ravioli find each source file based on file extensions, the user
specifies the files directly.

This option was added so that ravioli could be run as an automated step
as a post build step.